procesamiento_datos_almacenamiento_recuperacion_texto/requirements.txt
adlfs==2024.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.1
azure-data-tables==12.5.0
azure-datalake-store==0.0.53
azure-functions==1.19.0
azure-functions-durable==1.2.9
azure-identity==1.16.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
executing==2.0.1
frozenlist==1.4.1
fsspec==2024.3.1
furl==2.1.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.50
marshmallow==3.21.1
matplotlib-inline==0.1.7
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.23.6
orderedmultidict==1.0.1
orjson==3.10.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
pinecone-client==3.2.2
platformdirs==4.2.1
portalocker==2.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.6.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
